Square/requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
apimatic-core==0.1.4
apimatic-core-interfaces==0.1.7
apimatic-requests-client-adapter==0.1.3
attrs==25.4.0
blinker==1.9.0
CacheControl==0.12.14
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
deprecation==2.1.0
docutils==0.22.2
enum34==1.1.10
filetype==1.2.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jsonpickle==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-Garden==0.1.5
kivy_deps.sdl2==0.8.0
MarkupSafe==3.0.3
msgpack==1.1.2
narwhals==2.9.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.0
pyarrow==21.0.0
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.19.2
pypiwin32==223
python-dateutil==2.8.2
pytz==2025.2
pywin32==311
referencing==0.37.0
requests==2.28.2
rpds-py==0.28.0
setuptools==58.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
squareup==23.0.0.20221019
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
watchdog==6.0.0
wheel==0.45.1
